double

+++ jpsiphi/MuMuKKPAT/interface/MuMuProducerPAT.h
@@ -216,7 +216,7 @@ private:
 
   edm::InputTag inputGEN_;
   std::string vtxSample_;
-  std::vector<float> jspiMassCuts_,psiMassCuts_;
+  std::vector<double> jspiMassCuts_,psiMassCuts_;
   bool doData_, doMC_,addCommonVertex_,resolveAmbiguity_,addMuMulessPrimaryVertex_,addMCTruth_;
   int  MCParticle_;
   bool MCExclusiveDecay_;
